Reduce to minimal implementation
Shrink down the size of the user-space application by removing the
helper functions and directly using the system call functions.
Reason being that the inital plan was to test the character device
driver in an multi threaded environment, where helper functions would
have been easier to refactor, but I now keep a single thread environment.

Remove the open and close functions from the kernel module, as these are
not required either.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>